moved layout and structure out of doc-elements dir
